Fix OpCode reading on node packet callback handling
In jsimonetti/go-artnet#18 @h3ndrk introduced node OpCode callbacks, but unfortunatelly this broke the only packet which behaves completely different, which is ArtPollReply. And thus broke the controller logic.

I decided to add a method to the ArtNetPacket interface because we already have the (well tested) logic of decoding the OpCode inside of the packets validate method.

Signed-off-by: Alexander Pinnecke <alexander.pinnecke@googlemail.com>